skos2json/requirements.txt
sqlalchemy
rdflib
python-skos
rdflib-jsonld